feat: prevent deletion of users with active loans

+++ api/src/main/java/com/pachedev/library/service/UserService.java
@@ -5,15 +5,18 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.pachedev.library.model.User;
+import com.pachedev.library.repository.LoanRepository;
 import com.pachedev.library.repository.UserRepository;
 
 @Service
 public class UserService {
 
     private final UserRepository userRepository;
+    private final LoanRepository loanRepository;
 
-    public UserService(UserRepository userRepository) {
+    public UserService(UserRepository userRepository, LoanRepository loanRepository) {
         this.userRepository = userRepository;
+        this.loanRepository = loanRepository;
     }
 
     public User create(User newUser) {
@@ -45,6 +48,10 @@ public class UserService {
 
     public void delete(Long id) {
         User existingUser = findById(id);
+
+        if (loanRepository.existsByBookIdAndReturnDateIsNull(id)) {
+            throw new IllegalArgumentException("Cannot delete user with active loans");
+        }
         userRepository.delete(existingUser);
     }
 }
